Moves constants. Updates shadow declaration
